+++ CHANGES.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2025-10-09
+* Added `bw_timex.utils.interactive_td_widget` for drafting and visualizing temporal distributions in jupyter notebooks
+* Improved caching for activity name lookup
+* Improved logging
+
 ## [0.3.1] - 2025-03-31
 * Fixed an issue with non-unitary production exchanges